python: split generate_solver
